refactor(ui): partially flatten `workflows/components/` dir
- where there was only 1 or 2 files per folder, flatten them
- the rest I need to take a harder look at
- and this directory _does_ have too many files in right now
  - I would _probably_ split this into directories by the WorkflowDetails page and the WorkflowsList page
    - i.e. the same way it is split visually and in the routing, for intuitivity
