+++ ui/src/app/workflows/components/workflow-details/workflow-details.tsx
@@ -32,12 +32,12 @@ import {ResubmitWorkflowPanel} from '../resubmit-workflow-panel';
 import {RetryWorkflowPanel} from '../retry-workflow-panel';
 import {WorkflowArtifacts} from '../workflow-artifacts';
 import {WorkflowLogsViewer} from '../workflow-logs-viewer/workflow-logs-viewer';
-import {WorkflowNodeInfo} from '../workflow-node-info/workflow-node-info';
-import {WorkflowPanel} from '../workflow-panel/workflow-panel';
+import {WorkflowNodeInfo} from '../workflow-node-info';
+import {WorkflowPanel} from '../workflow-panel';
 import {WorkflowParametersPanel} from '../workflow-parameters-panel';
 import {WorkflowSummaryPanel} from '../workflow-summary-panel';
-import {WorkflowTimeline} from '../workflow-timeline/workflow-timeline';
-import {WorkflowYamlViewer} from '../workflow-yaml-viewer/workflow-yaml-viewer';
+import {WorkflowTimeline} from '../workflow-timeline';
+import {WorkflowYamlViewer} from '../workflow-yaml-viewer';
 import {ArtifactPanel} from './artifact-panel';
 import {SuspendInputs} from './suspend-inputs';
 import {WorkflowResourcePanel} from './workflow-resource-panel';
