+++ ui/src/app/workflows/components/workflow-drawer.scss
