+++ ui/src/app/workflows/components/workflow-from.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {labels} from '../../../models';
 import {ClusterWorkflowTemplateLink} from '../../cluster-workflow-templates/components/cluster-workflow-template-link';
-import {CronWorkflowLink} from '../../cron-workflows/components/cron-workflow-link';
-import {WorkflowTemplateLink} from '../../workflow-templates/components/workflow-template-link';
+import {CronWorkflowLink} from '../../cron-workflows/cron-workflow-link';
+import {WorkflowTemplateLink} from '../../workflow-templates/workflow-template-link';
 
 export function WorkflowFrom(props: {namespace: string; labels: {[name: string]: string}}) {
     const workflowTemplate = props.labels[labels.workflowTemplate];
