+++ ui/src/app/workflows/components/workflow-labels.scss
