+++ ui/src/app/workflows/components/workflow-logs-viewer/json-logs-field-selector.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import {useState} from 'react';
+
 import {TextInput} from '../../../shared/components/text-input';
 
 export interface SelectedJsonFields {
