+++ ui/src/app/workflows/components/workflow-logs-viewer/workflow-logs-viewer.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
-
 import {Autocomplete} from 'argo-ui';
 import {Observable} from 'rxjs';
 import {map, publishReplay, refCount} from 'rxjs/operators';
+
 import * as models from '../../../../models';
 import {execSpec} from '../../../../models';
 import {ANNOTATION_KEY_POD_NAME_VERSION} from '../../../shared/annotations';
