+++ ui/src/app/workflows/components/workflow-node-info.scss
