+++ ui/src/app/workflows/components/workflows-container.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
 import {WorkflowDetails} from './workflow-details/workflow-details';
-import {WorkflowsList} from './workflows-list/workflows-list';
+import {WorkflowsList} from './workflows-list';
 
 export const WorkflowsContainer = (props: RouteComponentProps<any>) => (
     <Switch>
