+++ ui/src/app/workflows/components/workflows-list.scss
