+++ ui/src/app/workflows/components/workflows-list.tsx
@@ -3,29 +3,29 @@ import * as React from 'react';
 import {useContext, useEffect, useMemo, useState} from 'react';
 import {RouteComponentProps} from 'react-router-dom';
 
-import * as models from '../../../../models';
-import {isArchivedWorkflow, Workflow, WorkflowPhase, WorkflowPhases} from '../../../../models';
-import {uiUrl} from '../../../shared/base';
-import {CostOptimisationNudge} from '../../../shared/components/cost-optimisation-nudge';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {ExampleManifests} from '../../../shared/components/example-manifests';
-import {Loading} from '../../../shared/components/loading';
-import {PaginationPanel} from '../../../shared/components/pagination-panel';
-import {useCollectEvent} from '../../../shared/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {historyUrl} from '../../../shared/history';
-import {ListWatch, sortByYouth} from '../../../shared/list-watch';
-import {Pagination, parseLimit} from '../../../shared/pagination';
-import {ScopedLocalStorage} from '../../../shared/scoped-local-storage';
-import {services} from '../../../shared/services';
-import {Utils} from '../../../shared/utils';
-import * as Actions from '../../../shared/workflow-operations-map';
-import {WorkflowCreator} from '../workflow-creator';
-import {WorkflowFilters} from '../workflow-filters/workflow-filters';
-import {WorkflowsRow} from '../workflows-row/workflows-row';
-import {WorkflowsSummaryContainer} from '../workflows-summary-container/workflows-summary-container';
-import {WorkflowsToolbar} from '../workflows-toolbar/workflows-toolbar';
+import * as models from '../../../models';
+import {isArchivedWorkflow, Workflow, WorkflowPhase, WorkflowPhases} from '../../../models';
+import {uiUrl} from '../../shared/base';
+import {CostOptimisationNudge} from '../../shared/components/cost-optimisation-nudge';
+import {ErrorNotice} from '../../shared/components/error-notice';
+import {ExampleManifests} from '../../shared/components/example-manifests';
+import {Loading} from '../../shared/components/loading';
+import {PaginationPanel} from '../../shared/components/pagination-panel';
+import {useCollectEvent} from '../../shared/use-collect-event';
+import {ZeroState} from '../../shared/components/zero-state';
+import {Context} from '../../shared/context';
+import {historyUrl} from '../../shared/history';
+import {ListWatch, sortByYouth} from '../../shared/list-watch';
+import {Pagination, parseLimit} from '../../shared/pagination';
+import {ScopedLocalStorage} from '../../shared/scoped-local-storage';
+import {services} from '../../shared/services';
+import {Utils} from '../../shared/utils';
+import * as Actions from '../../shared/workflow-operations-map';
+import {WorkflowCreator} from './workflow-creator';
+import {WorkflowFilters} from './workflow-filters';
+import {WorkflowsRow} from './workflows-row/workflows-row';
+import {WorkflowsSummaryContainer} from './workflows-summary-container';
+import {WorkflowsToolbar} from './workflows-toolbar';
 
 import './workflows-list.scss';
