+++ ui/src/app/workflows/components/workflows-row/workflows-row-name.tsx
@@ -2,11 +2,12 @@ import * as kubernetes from 'argo-ui/src/models/kubernetes';
 import * as React from 'react';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
+
 import {ANNOTATION_DESCRIPTION, ANNOTATION_TITLE} from '../../../shared/annotations';
 
 require('./workflows-row.scss');
 
-export const WorkflowsRowName = ({metadata}: {metadata: kubernetes.ObjectMeta}) => {
+export function WorkflowsRowName({metadata}: {metadata: kubernetes.ObjectMeta}) {
     const title = (metadata.annotations && metadata.annotations[ANNOTATION_TITLE]) || metadata.name;
     const description = (metadata.annotations && metadata.annotations[ANNOTATION_DESCRIPTION] && `\n${metadata.annotations[ANNOTATION_DESCRIPTION]}`) || '';
     const markdown = `${title}${description}`;
@@ -17,4 +18,4 @@ export const WorkflowsRowName = ({metadata}: {metadata: kubernetes.ObjectMeta}) 
             </ReactMarkdown>
         </div>
     );
-};
+}
