+++ ui/src/app/workflows/components/workflows-toolbar.tsx
@@ -2,11 +2,11 @@ import {NotificationType} from 'argo-ui';
 import * as React from 'react';
 import {useContext, useMemo} from 'react';
 
-import {isArchivedWorkflow, isWorkflowInCluster, Workflow} from '../../../../models';
-import {Context} from '../../../shared/context';
-import {services} from '../../../shared/services';
-import * as Actions from '../../../shared/workflow-operations-map';
-import {WorkflowOperation, WorkflowOperationAction, WorkflowOperationName} from '../../../shared/workflow-operations-map';
+import {isArchivedWorkflow, isWorkflowInCluster, Workflow} from '../../../models';
+import {Context} from '../../shared/context';
+import {services} from '../../shared/services';
+import * as Actions from '../../shared/workflow-operations-map';
+import {WorkflowOperation, WorkflowOperationAction, WorkflowOperationName} from '../../shared/workflow-operations-map';
 
 import './workflows-toolbar.scss';
